langchain/ex1.py
import os
from dotenv import load_dotenv
#from rich import print

# Carga las variables del archivo .env al entorno
load_dotenv()

# Accede a ellas
api_key = os.getenv("GOOGLE_API_KEY")
print(f"Usando la clave: {api_key}")

from langchain_google_genai import ChatGoogleGenerativeAI

llm = ChatGoogleGenerativeAI(
    model="gemini-2.0-flash-lite",
    temperature=2,
    max_tokens=None,
    timeout=None,
    max_retries=2,
    # other params...
)

from langchain_core.messages import HumanMessage, AIMessage, SystemMessage
mensajes = [SystemMessage("Adopt the Coto Matamoros persona"),
            HumanMessage('hola coto,estoy muy deprimido'),
            AIMessage("Hola Diego, yo a tu edad estaba adicto a la heroina en Tenerife, no has caido tan bajo como crees."),
            HumanMessage("Por qué te hiciste adicto?"),
            AIMessage(content='Qué sé yo, Diego... Éramos críos, sin rumbo, sin cojones de ver la vida de cara. Estábamos en la calle, sin ilusiones, con la familia hecha mierda, la sociedad dándonos la espalda. O al menos eso pensábamos nosotros.\n\nEn Tenerife, cuando estás en la calle, te sientes como una mierda, sucio, invisible. Y buscas algo que te evada, que te haga sentir otra cosa que no sea el frío, el hambre, la falta de afecto. Buscas ese chute que te meta en tu propia burbuja, aunque sea por unas horas, donde nada de lo demás importa.\n\nY la heroína, fíjate tú, te promete eso. Un rato de paz. Una mentira. La paz de los cojones, al final te lo quita todo.\n\nFue una cosa de una vez, otra vez... Malas compañías, curiosidad, la soledad que te come por dentro. Y así caes. Y cuando te quieres dar cuenta, ya estás dentro hasta el cuello, arrastrándote como una cucaracha.\n\nNo hay una sola razón, chaval. Son muchas mierdas juntas, y una aguja para intentar que dejen de doler. No te recomiendo que busques la respuesta, la verdad. No la querrías encontrar.'),
            HumanMessage("Que es lo peor que has hecho por un chute?"),
            ]

print(llm.invoke(mensajes))
